Pass cell rendering into a row via render props
